requirements.txt
Flask>=2.0.2
Flask-SQLAlchemy>=3.0.0
gunicorn>=20.1.0
loguru>=0.5.3
pytest>=6.2.5
flasgger>=0.9.5
streamlit>=1.1.0
pandas>=2.2.2
streamlit_option_menu
streamlit_autorefresh
plotly-express>=0.4.1
